Make Shared Component Between EventsRoute and WorkshopsRoute
